unit-test and sonarLint

+++ src/main/java/com/musala/task/drone/entity/Medication.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import javax.persistence.*;
 import javax.validation.constraints.Pattern;
-import java.time.LocalDateTime;
 
 @Data
 @Entity
@@ -13,28 +12,28 @@ public class Medication {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name="id")
+    @Column(name = "id")
     private Integer id;
 
-    @Column(name = "code",nullable = false)
-    @Pattern(regexp = "^[A-Z0-9_]+$",message = "allowed only upper case letters, underscore and numbers")
+    @Column(name = "code", nullable = false)
+    @Pattern(regexp = "^[A-Z0-9_]+$", message = "allowed only upper case letters, underscore and numbers")
 
 
     private String code;
 
-    @Column(name = "name",nullable = false)
-    @Pattern(regexp = "^[a-zA-Z0-9_-]+$",message = "allowed only letters, numbers, ‘-‘, ‘_’")
+    @Column(name = "name", nullable = false)
+    @Pattern(regexp = "^[a-zA-Z0-9_-]+$", message = "allowed only letters, numbers, ‘-‘, ‘_’")
     private String name;
 
-    @Column(name = "weight",nullable = false)
-    private double weight;
+    @Column(name = "weight", nullable = false)
+    private Integer weight;
 
-    @Column(name = "image",nullable = false)
+    @Column(name = "image", nullable = false)
     @Lob
     private byte[] image;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "drone_id",referencedColumnName = "id",nullable = false)
+    @JoinColumn(name = "drone_id", referencedColumnName = "id", nullable = false)
     private Drone drone;
 
 }
